SCREEN SHAAAAAAAKE and linear knockback implemented
Also corrected moving whle attacking in the air to be more of a "hover", and attacking up doesn't move the character at all
slightly modified jumping off ledges to give more limited control to the jump, and justify the climb.
also knockback falls off, the longer it takes for something to get hit the less knockback it takes.
also added that cape i made