further streamlined making attacks, expanded the bash dash functionality
created a hitNode array which makes it very easy to create attacks with multiple hitboxes - high potential for expanded combo moves
attacks which follow the player and attacks which do not made more consistent and implementing them has been significantly streamlined
the bash dash has been implemented such that if a wall is hit, or an enemy which did not die, the player is bounced back. think I'll also trigger the attack_done function when this happens as well.