Slight combotimer bugfix, further improved attack instancing logic, created quick slash attacks
The combotimer being set at the start of attacks was logically inconsistent. The timer is now *stopped* at the start of attacks and started at the end of attacks.
Attack instancing logic now works off the last hitbox parent nodes added to the array in each function. this allows the spawning of several distinct attacks in one attack animation.
Quick attacks implemented to a satisfactory degree. Some animation changes could help it to feel better.